Added complete factories with a matcher for verification, #158
Complete factories are factories that should include all attributes and
associations possible for a model. The matcher verifies this by checking
that all not excluded attributes and associations are present on a
model.

The reason for these factories is to verify new functionality is not
forgotten in other parts of the code. The main reason is the creation of
suites and evaluation from templates, where new functionality on either
suites or evaluations must be handled when creating from templates.