Fixed problem with duplicates suites, #203
Overlapping roles (suite_member and
suite_manager) that result in duplicates suites in the suites listing
for some users. This fix adds a contraint to only fetch unqieue suites.

The reason why some user get muliple and overlapping roles still needs to be
addressed for verifying that this behaviour is OK.